requirements.txt
alembic==1.13.2
bcrypt==4.2.0
bidict==0.23.1
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
geographiclib==2.0
geopy==2.4.1
greenlet==3.0.3
h11==0.14.0
idna==3.8
itsdangerous==2.2.0
Jinja2==3.1.4
Mako==1.3.5
MarkupSafe==2.1.5
pillow==10.4.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
python-engineio==4.9.1
python-socketio==5.11.3
requests==2.32.3
simple-websocket==1.0.0
SQLAlchemy==2.0.32
typing_extensions==4.12.2
urllib3==2.2.2
uWSGI==2.0.26
websocket-client==1.8.0
Werkzeug==3.0.3
wsproto==1.2.0
WTForms==3.1.2
